# Conflicts:
#	.github/upstream-workflows-disabled/build-and-upload-images.yml
#	.github/upstream-workflows-disabled/periodic-test.yml
#	.github/upstream-workflows-disabled/pr-no-generated-changes.yml
#	.github/upstream-workflows-disabled/publish.yml
#	.github/upstream-workflows-disabled/release-please.yml
#	.tool-versions
#	docs/ARCHITECTURE.md
#	packages/api/internal/api/api.gen.go
#	packages/api/internal/orchestrator/create_instance.go
#	packages/api/internal/orchestrator/nodemanager/mock.go
#	packages/api/internal/orchestrator/nodemanager/reachability_test.go
#	packages/api/internal/orchestrator/nodemanager/status.go
#	packages/api/internal/orchestrator/nodemanager/sync.go
#	packages/api/internal/orchestrator/placement/errors.go
#	packages/api/internal/orchestrator/placement/errors_test.go
#	packages/db/pkg/testutils/db.go
#	packages/docker-reverse-proxy/go.mod
#	packages/docker-reverse-proxy/go.sum
#	packages/docker-reverse-proxy/internal/cache/auth.go
#	packages/docker-reverse-proxy/internal/constants/main.go
#	packages/docker-reverse-proxy/internal/handlers/proxy.go
#	packages/docker-reverse-proxy/internal/handlers/store.go
#	packages/docker-reverse-proxy/internal/handlers/token.go
#	packages/docker-reverse-proxy/main.go
#	packages/envd/go.mod
#	packages/envd/internal/port/forward_test.go
#	packages/envd/internal/port/scan_subscriber.go
#	packages/envd/pkg/version.go
#	packages/envd/routes_test.go
#	packages/envd/spec/envd.yaml
#	packages/orchestrator/Makefile
#	packages/orchestrator/cmd/smoketest/smoke_test.go
#	packages/orchestrator/pkg/sandbox/network/slot.go
#	packages/orchestrator/pkg/sandbox/network/storage_kv.go
#	packages/orchestrator/pkg/sandbox/network/storage_local.go
#	packages/orchestrator/pkg/sandbox/network/storage_memory.go
#	packages/orchestrator/pkg/sandbox/reboot.go
#	packages/orchestrator/pkg/sandbox/reclaim.go
#	packages/orchestrator/pkg/server/prefetch_harvest.go
#	packages/orchestrator/pkg/template/build/phases/finalize/configure_test.go
#	packages/shared/go.mod
#	packages/shared/pkg/featureflags/flags.go
#	packages/shared/pkg/sandbox-network/firewall_test.go
#	packages/shared/pkg/telemetry/meters.go

+++ packages/db/pkg/dashboard/queries/models.go
@@ -30,12 +30,12 @@ type Team struct {
 type TeamLimit struct {
 	ID                       uuid.UUID
 	MaxLengthHours           int64
-	ConcurrentSandboxes      int32
-	ConcurrentTemplateBuilds int32
-	MaxVcpu                  int32
-	MaxRamMb                 int32
-	DiskMb                   int32
-	EventsTtlDays            int32
+	ConcurrentSandboxes      int64
+	ConcurrentTemplateBuilds int64
+	MaxVcpu                  int64
+	MaxRamMb                 int64
+	DiskMb                   int64
+	EventsTtlDays            int64
 	DefaultFreeDiskSizeMb    int64
 	MaxDiskSizeMb            int64
 }
